refactor(simulation): extract analysis helpers
